Implement OP_CONSTANT_LONG instruction.
The instruction is similar to OP_CONSTANT, but instead of a 8-bit
operand, it has a 24-bit operand. It's intended to make clox
support programs with more than 256 constants.

writeConstant() will automatically choose between OP_CONSTANT and
OP_CONSTANT_LONG based on the number of constants that have already
been stored in Chunk.